Set up separate Community page
Securing the home page messed up registration. It no longer showed the
notice about confirming your email. Now, the home page will be
unsecured, and there's a new community page that will be the "home" page
for signed-in users.
