+++ app/helpers/community_helper.rb
@@ -1,0 +1,2 @@
+module CommunityHelper
+end
