+++ app/views/community/_list.html.erb
