+++ app/views/community/show.html.erb
@@ -1,0 +1,20 @@
+<div>
+  <h1>Simply Always Awake Community Locator</h1>
+</div>
+
+<div>
+  <h1>My Community</h1>
+</div>
+
+<% if @community.empty? %>
+  <div>There are currently no people in your area.</div>
+<% else %>
+  <div>
+    <%= render 'list' %>
+  </div>
+<% end %>
+
+<%= link_to "Edit location", location_path, class: "rounded-lg m-2 py-3 ml-2 px-5 bg-gray-100 inline-block font-medium" %>
+<div>
+  <%= link_to 'Sign out', destroy_user_session_path, data: { "turbo-method": :delete }, class: "rounded-lg py-3 m-2 px-5 bg-gray-100 inline-block font-medium" %>
+</div>
